Finish game details api implementation

+++ data/src/main/java/com/github/data/model/GameDetailsResponse.kt
@@ -7,12 +7,12 @@ import com.google.gson.annotations.SerializedName
  * Contact: muhamed.gendy@gmail.com
  */
 data class GameDetailsResponse(
-  val id: Int?,
+  val id: Long?,
   val slug: String?,
   val name: String?,
   @SerializedName("name_original") val nameOriginal: String?,
   val description: String?,
-  val metacritic: Int?,
+  val metacritic: Long?,
   val released: String?,
   val tba: Boolean?,
   val updated: String?,
@@ -20,32 +20,32 @@ data class GameDetailsResponse(
   @SerializedName("background_image_additional") val backgroundImageAdditional: String?,
   val website: String?,
   val rating: Double?,
-  @SerializedName("rating_top") val ratingTop: Int?,
+  @SerializedName("rating_top") val ratingTop: Long?,
   val ratings: List<Rating>?,
-  val added: Int?,
-  val playtime: Int?,
-  @SerializedName("screenshots_count") val screenshotsCount: Int?,
-  @SerializedName("movies_count") val moviesCount: Int?,
-  @SerializedName("creators_count") val creatorsCount: Int?,
-  @SerializedName("achievements_count") val achievementsCount: Int?,
-  @SerializedName("parent_achievements_count") val parentAchievementsCount: Int?,
+  val added: Long?,
+  val playtime: Long?,
+  @SerializedName("screenshots_count") val screenshotsCount: Long?,
+  @SerializedName("movies_count") val moviesCount: Long?,
+  @SerializedName("creators_count") val creatorsCount: Long?,
+  @SerializedName("achievements_count") val achievementsCount: Long?,
+  @SerializedName("parent_achievements_count") val parentAchievementsCount: Long?,
   @SerializedName("reddit_url") val redditUrl: String?,
   @SerializedName("reddit_name") val redditName: String?,
   @SerializedName("reddit_description") val redditDescription: String?,
   @SerializedName("reddit_logo") val redditLogo: String?,
-  @SerializedName("reddit_count") val redditCount: Int?,
-  @SerializedName("twitch_count") val twitchCount: Int?,
-  @SerializedName("youtube_count") val youtubeCount: Int?,
-  @SerializedName("reviews_text_count") val reviewsTextCount: Int?,
-  @SerializedName("ratings_count") val ratingsCount: Int?,
-  @SerializedName("suggestions_count") val suggestionsCount: Int?,
+  @SerializedName("reddit_count") val redditCount: Long?,
+  @SerializedName("twitch_count") val twitchCount: Long?,
+  @SerializedName("youtube_count") val youtubeCount: Long?,
+  @SerializedName("reviews_text_count") val reviewsTextCount: Long?,
+  @SerializedName("ratings_count") val ratingsCount: Long?,
+  @SerializedName("suggestions_count") val suggestionsCount: Long?,
   @SerializedName("alternative_names") val alternativeNames: List<String>?,
   @SerializedName("metacritic_url") val metacriticUrl: String?,
-  @SerializedName("parents_count") val parentsCount: Int?,
-  @SerializedName("additions_count") val additionsCount: Int?,
-  @SerializedName("game_series_count") val gameSeriesCount: Int?,
+  @SerializedName("parents_count") val parentsCount: Long?,
+  @SerializedName("additions_count") val additionsCount: Long?,
+  @SerializedName("game_series_count") val gameSeriesCount: Long?,
   @SerializedName("user_game") val userGame: Any?,
-  @SerializedName("reviews_count") val reviewsCount: Int?,
+  @SerializedName("reviews_count") val reviewsCount: Long?,
   @SerializedName("saturated_color") val saturatedColor: String?,
   @SerializedName("dominant_color") val dominantColor: String?,
   val platforms: List<Platform>?,
@@ -58,9 +58,9 @@ data class GameDetailsResponse(
 ) {
 
   data class Rating(
-    val id: Int?,
+    val id: Long?,
     val title: String?,
-    val count: Int?,
+    val count: Long?,
     val percent: Double?
   )
 
@@ -70,13 +70,13 @@ data class GameDetailsResponse(
     val requirements: Requirements?
   ) {
     data class Platform(
-      val id: Int?,
+      val id: Long?,
       val name: String?,
       val slug: String?,
       val image: Any?,
       @SerializedName("year_end") val yearEnd: Any?,
       @SerializedName("year_start") val yearStart: Any?,
-      @SerializedName("games_count") val gamesCount: Int?,
+      @SerializedName("games_count") val gamesCount: Long?,
       @SerializedName("image_background") val imageBackground: String?
     )
 
@@ -87,41 +87,41 @@ data class GameDetailsResponse(
   }
 
   data class Store(
-    val id: Int?,
+    val id: Long?,
     val url: String?,
     val store: Store?
   ) {
     data class Store(
-      val id: Int?,
+      val id: Long?,
       val name: String?,
       val slug: String?,
       val domain: String?,
-      @SerializedName("games_count") val gamesCount: Int?,
+      @SerializedName("games_count") val gamesCount: Long?,
       @SerializedName("image_background") val imageBackground: String?
     )
   }
 
   data class Developer(
-    val id: Int?,
+    val id: Long?,
     val name: String?,
     val slug: String?,
-    @SerializedName("games_count") val gamesCount: Int?,
+    @SerializedName("games_count") val gamesCount: Long?,
     @SerializedName("image_background") val imageBackground: String?
   )
 
   data class Genre(
-    val id: Int?,
+    val id: Long?,
     val name: String?,
     val slug: String?,
-    @SerializedName("games_count") val gamesCount: Int?,
+    @SerializedName("games_count") val gamesCount: Long?,
     @SerializedName("image_background") val imageBackground: String?
   )
 
   data class Publisher(
-    val id: Int?,
+    val id: Long?,
     val name: String?,
     val slug: String?,
-    @SerializedName("games_count") val gamesCount: Int?,
+    @SerializedName("games_count") val gamesCount: Long?,
     @SerializedName("image_background") val imageBackground: String?
   )
 }
